pyproject.toml
[project]
name = "portfolio-profile-chatbot"
version = "0.1.0"
description = "Gradio + Gemini portfolio chatbot (iframe-ready)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "gradio>=5.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pymupdf>=1.24.0",
]

[tool.uv]
package = false
